requirements.txt
google-generativeai
python-dotenv
langchain
pypdf
chromadb
faiss-cpu
langchain_google_genai
fastapi
pydantic
uvicorn
requests